Move more PC methods that don't need the template parameters into base class. (#1760)
